Move main to cmd/ and handlers to handlers/
